requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.2.0
azure-core==1.26.4
azure-functions==1.15.0
azure-identity==1.14.0
azure-storage-blob==12.16.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Django==4.2.13
django-filter==24.2
django-silk==5.1.0
django-storages==1.13.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gprof2dot==2024.6.6
gremlinpython==3.7.2
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
kombu==5.3.7
Markdown==3.6
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
nest-asyncio==1.6.0
packaging==24.0
pillow==10.3.0
portalocker==2.8.2
prompt_toolkit==3.0.46
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
redis==5.0.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.19.0
